requirements.txt
numpy>=1.19.2
imlib>=0.0.27
opencv-python>=4.4.0.44
imutils>=0.5.3
pyyaml
playsound
